pyproject.toml
[tool.poetry]
name = "cookbook"
version = "0.1.0"
description = ""
authors = ["Anton Skvortsov <askvrtsv@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.1"
django-environ = "^0.11.2"
django-debug-toolbar = "^4.2.0"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.9"
gunicorn = "^21.2.0"
whitenoise = "^6.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
